server_deploy_backup_20260315/backup_deploy/integrity-tools/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.0.0
flask-cors>=4.0.0
pyjwt>=2.8.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
edge-tts>=6.1.0
requests>=2.31.0
pdfplumber>=0.10.0
Pillow>=10.0.0
pypdf>=3.0.0
gunicorn>=21.0.0
aiohttp>=3.9.0